Small fix to datatypes

+++ pkg/resampler/resampler.go
@@ -35,7 +35,7 @@ func SimpleResampler(data []util.Float, old_frequency int64, new_frequency int64
 
 // Inspired by one of the algorithms from https://skemman.is/bitstream/1946/15343/3/SS_MSthesis.pdf
 // Adapted from https://github.com/haoel/downsampling/blob/master/core/lttb.go
-func LargestTriangleThreeBucket(data []util.Float, old_frequency int, new_frequency int) ([]util.Float, int, error) {
+func LargestTriangleThreeBucket(data []util.Float, old_frequency int64, new_frequency int64) ([]util.Float, int64, error) {
 
 	if old_frequency == 0 || new_frequency == 0 {
 		return data, old_frequency, nil
